usbd_gs_can: clean up send to host path
The USBD_GS_CAN_SendToHost() function is used to send to struct
gs_host_frame_object to the host.

After the sending process has been started, the outgoing frame object
is added to the list of free objects and the variable
USBD_GS_CAN_HandleTypeDef::TxState is used to track if the transfer to
the host is in progress.

Instead hold the outgoing object in
USBD_GS_CAN_HandleTypeDef::to_host_buf and move it to the free list
after the transfer is finished in USBD_GS_CAN_DataIn(). Use this to
track if a transfer is ongoing.